Update remix starter
