+++ starters/remix-starter/app/db/execute-query.direct.ts
@@ -1,18 +1,40 @@
 import { DataAccessor } from '@/lib/data-access/dataAccessor';
+import { decodeUrlCompletely } from '@/utils/url-decoder';
 
-let databaseUrlDecoded = process.env.CONNECTION_URL || '';
+export type QueryData<T> =
+  | { isError: false; data: T }
+  | {
+      isError: true;
+      errorMessage: string;
+    };
 
-// Handle double URL encoding - decode until we get a valid URL format
-while (databaseUrlDecoded.includes('%')) {
-  const decoded = decodeURIComponent(databaseUrlDecoded);
-  if (decoded === databaseUrlDecoded) break; // No more decoding needed
-  databaseUrlDecoded = decoded;
+export async function executeQuery<T>(query: string, params?: string[]): Promise<QueryData<T[]>> {
+  let result: { data: T[] };
+
+  try {
+    result = await executeQueryDirectly<T>(query, params);
+
+    return {
+      ...result,
+      isError: false,
+    };
+  } catch (error) {
+    const typedError = error as Error;
+    const errorMessage = typedError?.message || `Something went wrong executing the query: ${query}`;
+
+    return {
+      isError: true,
+      errorMessage,
+    };
+  }
 }
-let accessor: ReturnType<typeof DataAccessor.getAccessor> | null = null;
+
+let accessor: ReturnType<typeof DataAccessor.getDatabaseAccessor> | null = null;
 
 async function getAccessor() {
   if (!accessor) {
-    accessor = DataAccessor.getAccessor(databaseUrlDecoded);
+    accessor = DataAccessor.getDatabaseAccessor(process.env.DATA_SOURCE_TYPE);
+    const databaseUrlDecoded = decodeUrlCompletely(process.env.CONNECTION_URL || '');
     await accessor.initialize(databaseUrlDecoded);
   }
